tst-syscall*: Add tests for syscall() with varargs
Add tests in preparation for genericizing some of the architecture
syscall() implementations.

This was noticed when testing OR1K and found it had a broken syscall
implementation.

These tests try to cover the libc syscall() lqyer which has the purpose
of passing the syscall number and arguments to the kernel.  The actual
kernel syscalls chosen have been selected for ease of testing.

Signed-off-by: Stafford Horne <shorne@gmail.com>